% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{generate_dependency}
\alias{generate_dependency}
\title{Generate dependency}
\usage{
generate_dependency(file)
}
\arguments{
\item{file}{R file from which the CSS is generated.}
}
\description{
Generate dependency
}
