% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roclet_style}
\alias{roclet_style}
\title{Roclets}
\usage{
roclet_style()
}
\description{
These functions implement \code{\link[roxygen2:roclet]{roxygen2::roclet()}}s for integration
with the roxygen2 package.
}
