% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{style}
\alias{style}
\title{Style}
\usage{
style(dir = NULL, file = NULL, out = "style.min.css", verbose = FALSE)
}
\arguments{
\item{dir}{Directory containing R files to scan.}

\item{file}{File to scan.}

\item{out}{Path to output CSS file.}

\item{verbose}{Whether to warn on classes that fail to parse.}
}
\description{
Generate CSS, set either \code{dir} or \code{file}.
}
