% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{style}
\alias{style}
\title{Style}
\usage{
style(dir, file, warn = FALSE)
}
\arguments{
\item{dir}{Directory containing R files to scan.}

\item{file}{Output CSS file.}

\item{warn}{Whether to warn on classes that fail to parse.}
}
\description{
Generate CSS
}
